Narrow the singular builder disables to the tests that actually fail
UserBuilderSingularTest and UserSuperSingularTest were forked: the source
suite ran them, the ASM suite carried a copy with a bare class-level
@Disabled and no reason, and the JDK suite inherited that copy. Fourteen
tests were skipped on the bytecode backends.

Only ten of them fail. The five singular-map tests in each class fail
because the bytecode backends call Map.put through the descriptor of
Collection.add:

  NoSuchMethodError: 'boolean java.util.LinkedHashMap.put(String, Integer)'

Keep one copy of each test in the TCK and disable just those five per class
for the bytecode backends, naming the reason, so the gap is visible and the
remaining four tests run on every backend instead of being skipped with the
rest. The fixture keeps the ASM suite's copy, whose source-suite twin had an
unused SuperBuilder import.

The underlying Map.put descriptor bug is left for a separate change.

All three suites run the same 432 tests as before; eight that were skipped
now run and pass.
