+++ test-suite-tck/src/main/java/io/micronaut/sourcegen/example/UserBuilderSingular.java
